Add files via upload
The `PythArdu-DataStringFilterTrial.rar` contains three files. A .ui made in QT, the converted .py file and an arduino code. The arduino code was tested on a arduino nano. One would need to modify the `COM8` in the python code with the one where the users arduino is connected. Please open the .ui file in PyQT to understand the object referencing as it would make reading the python code easier.